backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1

SQLAlchemy==2.0.29
alembic==1.13.1
PyMySQL==1.1.1

PyJWT==2.8.0
requests==2.32.3
Werkzeug==3.0.1
redis==5.0.1
python-socketio==5.11.1
eventlet==0.33.3

pydantic==2.10.4
pydantic-settings==2.1.0
pydantic[email]==2.10.4

google-genai